VisVector Fixes
"Wind" vector is no longer a separate vector from the gap, but instead a scalar multiple of the MinVectorSize (what was once called the gapVector).
This means that puzzles in 04 can still be dynamicly generated, but correct issues with player in a vector that doesn't seem to align with the goal post, even if the answer is correct.

Also refactored some code, did some clean up of unused functions and variables.  Still more to remove.  Want to try and simplify the code for the next team.