Stage04 - Level 2
New Level Created using the Stage04 mechanics
Stage04-1 has had the difficulty reduced to "1", meaning Stage04-1 is only for 1D vectors (X or Y vectors)
Stage04-2 is a new level design with a difficulty of "2", meaning Stage04-2 is for 2D vectors.

Updated the EnterExitSelect script to include the new level.

Minor tweaks to the Puzzle04Window script to fix a bug with UI cards adding twice and crashing if you complete (or exit) either stage04 level 1 or 2 and then enter the advanced difficulty.

Changes to VisualVector, Wind now dynmaicaly creates vectors as scalars (fractional or not) of the GapVector (now renamed MinVectorSize).