Rotateable and Zooming Camera in Puzzles
Like it says.  Now you can rotate the camera around in the puzzle.  The camera stays focused on the center of the puzzle (specifically halfway between the portal and the goal).

Hold Right Click and drag the mouse to rotate around the puzzle
Use the scroll wheel to zoom in and out (not really a zoom, depth of field isn't changing, the position of the camera is, w/e, works).

Added some accessibility changes to the grid under the properties in VisualVectors for changing the bar thickness and the node sizes.
Working on color changes.

Also dropped the difficulty on some of the puzzles to make it easier to do.  This may need more tweeaking.