Bug Fix: Stage04-1 Locking Up
There was a bug causing Stage04-1 to lock up.   The problem was requiring answer cards to be unique AND not a scalar in 1D space.  This meant that it was possible to be in an infinite loop and lock up the program.

There are only so many numbers to use and if you eliminate scalars as possibilities there are drastically less.  Especially in 1D where one card may generate "1" as an answer.  Now no cards that are a multiple of "1" are possible, which when concerning INTEGERS is impossible.