VisVector Update
Changes to VisualVector and associated UI changes to Puzzle04Window

VisVector allows a player to project their answer vectors onto the planes, this new change allows the player to turn on and off certain planes for better visibility.

Along with the toggling of projections, anchors associated with that plane will also toggle off.

Changes to how the toggling system works.  Simplified code for many of the toggles to just a simple toggle""() method rather than turnOn()/turnOFF() methods.