Pause Menu Extra Info and Keyboard Controls for Puzzle04
Added Keyboard controls for moving the camera when inside of a puzzle.  They follow the same controls as the character camera controller (I,J,K,L) plus the addition of U,O to zoom in and out.

Also added keyboard controls for increasing/decreasing slider 1/2 in the puzzle.  The player may use 1/2 or 3/4 in leu of sliding with them mouse.

All these new controls that are puzzle specific would be clunky on the pause menu's "Controls" element.  So a new puzzle specific controls window was added to the pause menu.

The new window is only active for Puzzle04 at the moment.  But the controls are simple enough.  If you want to activate it as an element for a puzzle, then that Puzzle's window controller needs to access the PauseWindow.cs.  And the PuzzleSystem.cs need to turn on the extra window when loading the scene.
After Puzzle##Window has the PauseWindow as an element, you can alter the text of the window with (your PauseWindow variable name).setPuzzleControlsText(string new Controls).  If the list of new controls is too long, cuts off, hard to read, you can resize that window with  (your PauseWindow variable name).resizePuzzleControls(float width, float height)

Also fixed a bug in the pause menu if you opened the settings menu while paused, and then exited the pause menu without exiting settings.  Before when you reopened the pause menu, the settings menu was still active.  That has been corrected.  If you close the pause menu, no matter what window was on, when you rePause it, the starting pause menu appears.