Fixes to Puzzle04-1 and Improved Grapple
Raised the cliffs and differences on Level 1, that should fix the camera sticking in the wrong spot when exiting a puzzle.  I didn't have it happen again

Grapple now fires even when a wrong answer is submitted.  The grapple fires in the direction of the answer vector when wrong, and in the direction of the goal when right.

Grapple now plays a sound when firing, and sound when it hits and a sound when it misses.  Also added some smoke effect so it's like firing a musket.