Puzzle04 Camera Changes and Axis Labels
Moved the cameras on Puzzle04-3 so that the cameras are closer to the player's starting position.

Added Axis Labels to the VisualVectors script.  The labels turn on and off with the Axis Toggle.
Axis labels are placed along their respective axis at a magnitude of 2 along their axis:
- Xaxis label is at (2,0,0), Y is at (0,2,0), Z is at (0,0,2)
Axis Labels only show up relevant to the puzzle answer ie: if the puzzle answer X is positive, then "+ X" shows up on the X axis at (2,0,0).  If the answer X is negative then "- X" is at (-2,0,0)
Axis labels always turn to face the camera as the player rotates around the puzzle