ver1.2a Database Implimentation, Admin Accounts, Progress Saving
Newest build includes some changes to database accounts and tables
Player data is now tracked and entered into the database for Stage completion times (if completed).  Some Stages have tracking for specific per puzzle data (not quite optimized for parsing, but it's in there).

Admin accounts are now a thing.  After an account is created in-game.  That account can be flagged on the database side as an Admin account.  After that subsequent logins using that account will trigger the Admin Options page before logging into the game.  Admins can create sessions and set their expiration time.  Later changes planned to allow Admins to create Sessions that lock players logging into that session into specific stages, or set joining players starting stats.  Code has to be refactored and unified a lot more before that can happen.

What is possible is any user can now complete a stage and have that stage data saved to the database.  When the player next logs in, the game will remember what stages they've completed and in what time they completed them in.  This allows the player to save their progress on play throughs.  More planned with this in the future.