Visual Changes and Ship Repair
Added some visual elements to certain levels.  Visual Vectors in Level 4 now has a particle effect in the direction of the vector to better tell the difference when the vectors overlap.  Also removed the static solid endpoints.

Ship Repair now updates as each puzzle is completed.  If it is the first time that puzzle has been completed there is a small animation takeover that shows what part was unlocked.  Subsequent logins do not play this animation, though when the player logs in, any parts that were unlocked remain unlocked and will display in the hubworld.

Changes to movement speed on the hubworld.  Increased the walking speed, dumbed down the running speed and increased gravity.  Trying to stop the player from rocketing off any small ramp if the player is holding shift.

UI changes to login.  Logging in now displays a little status animation when connecting to the database.  It relays to the player what action the game is taking when logging in and some useful tips while they wait.